requirements.txt
kivy==2.0.0
pymongo==4.1.1
werkzeug==2.1.2
dnspython~=2.2.1